requirements.txt
crewai[tools]
langchain
langchain-community
openai
python-dotenv
click
shutil